# dsh/dsh.openapi.yaml
openapi: 3.1.0
info:
  title: BThwani DSH API
  version: 0.3.0
  description: API contract for client-facing store discovery, catalog management, and checkout flow.
  contact:
    name: BThwani DSH API Governance
    url: https://bthwani.internal/governance/dsh
servers:
  - url: https://api.bthwani.internal/dsh
    description: Production DSH API boundary
  - url: http://localhost:8080
    description: Local DSH backend runtime
...
paths:
...
  /stores/{store_id}/categories:
    post:
      tags:
        - Category Structure
      summary: Create category for a store
      description: Creates a new category (and optional subcategory via parent_id) for the given store.
      operationId: createCategory
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Category Structure
      summary: List categories for a store
      description: Returns all categories for the given store.
      operationId: listCategories
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Categories list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCategoriesResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/{id}:
    get:
      tags:
        - Category Structure
      summary: Get a single category
      description: Returns one store category by ID, including hierarchy metadata used by DSH catalog surfaces.
      operationId: getCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRecord'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Category Structure
      summary: Update category
      description: Updates category details (name, parent_id, description).
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Updated category record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Category Structure
      summary: Delete category
      description: Deletes the category. Products assigned to this category will have their category_id set to null.
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category deleted
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── J-002 / DSH-SLICE-002C: Product Media ──────────────────────────────────
...
  /catalog-conflicts:
    get:
      tags:
        - Catalog Conflict Audit
      summary: List catalog overrides conflicts
      description: Returns a list of pending or resolved conflicts between central catalog and partner local overrides.
      operationId: listConflicts
      parameters:
        - name: store_id
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, resolved_accept_local, resolved_reverted]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Conflicts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListConflictsResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /catalog-conflicts/{id}/resolve:
    post:
      tags:
        - Catalog Conflict Audit
      summary: Resolve a catalog override conflict
      description: Resolves a catalog override conflict by either accepting the partner local override or reverting it to central.
      operationId: resolveConflict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveConflictRequest'
      responses:
        '200':
          description: Conflict resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveConflictResponse'
        '400':
          description: Validation or parameter error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Conflict not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── J-001 / Store Discovery ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional bearer token for personalized store discovery (e.g. retrieving user's favorite stores).
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    # ─── J-002 / DSH-SLICE-002A: Product Identity ─────────────────────────────
...
    CategoryRecord:
      type: object
      required:
        - id
        - store_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
        store_id:
          type: string
        parent_id:
          type: string
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        parent_id:
          type: string
        name:
          type: string
        description:
          type: string
    UpdateCategoryRequest:
      type: object
      properties:
        parent_id:
          type: string
        name:
          type: string
        description:
          type: string
    ListCategoriesResponse:
      type: object
      required:
        - categories
        - pagination
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ─── J-002 / DSH-SLICE-002C: Product Media ──────────────────────────────────
...
    CatalogConflict:
      type: object
      required:
        - id
        - store_id
        - product_id
        - product_name
        - conflict_type
        - central_value
        - override_value
        - status
        - created_at
      properties:
        id:
          type: string
        store_id:
          type: string
        product_id:
          type: string
        product_name:
          type: string
        conflict_type:
          type: string
          enum: [price_divergence, availability_divergence]
        central_value:
          type: string
        override_value:
          type: string
        status:
          type: string
          enum: [pending, resolved_accept_local, resolved_reverted]
        resolved_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    ResolveConflictRequest:
      type: object
      required:
        - resolution
      properties:
        resolution:
          type: string
          enum: [accept_local, revert_to_central]
          description: How to resolve the conflict.
    ResolveConflictResponse:
      type: object
      required:
        - conflict_id
        - status
      properties:
        conflict_id:
          type: string
        status:
          type: string
          enum: [resolved_accept_local, resolved_reverted]
    ListConflictsResponse:
      type: object
      required:
        - conflicts
        - limit
        - offset
        - total
      properties:
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/CatalogConflict'
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
...
    Pagination:
      type: object
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    # ─── J-003 / Checkout Schemas ─────────────────────────────────────────────
...